fix: move getFormFieldIconsPresets.tsx to src/testing
