+++ packages/vkui/src/components/Input/Input.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react';
-import { getFormFieldIconsPresets } from '../../lib/presets/getFormFieldIconsPresets';
 import { CanvasFullLayout, DisableCartesianParam } from '../../storybook/constants';
+import { getFormFieldIconsPresets } from '../../testing/getFormFieldIconsPresets';
 import { Input, InputProps } from './Input';
 
 const iconsPresets = getFormFieldIconsPresets();
