+++ packages/vkui/src/components/NativeSelect/NativeSelect.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react';
-import { getFormFieldIconsPresets } from '../../lib/presets/getFormFieldIconsPresets';
 import { CanvasFullLayout, DisableCartesianParam } from '../../storybook/constants';
+import { getFormFieldIconsPresets } from '../../testing/getFormFieldIconsPresets';
 import { NativeSelect, NativeSelectProps } from './NativeSelect';
 
 const story: Meta<NativeSelectProps> = {
