+++ packages/vkui/src/testing/getFormFieldIconsPresets.tsx
@@ -1,7 +1,7 @@
 import { Icon16Clear, Icon28MessageOutline } from '@vkontakte/icons';
 import { noop } from '@vkontakte/vkjs';
-import { IconButton } from '../../components/IconButton/IconButton';
-import { createFieldWithPresets } from '../../testing/createFieldWithPresets';
+import { IconButton } from '../components/IconButton/IconButton';
+import { createFieldWithPresets } from './createFieldWithPresets';
 
 export const getFormFieldIconsPresets = () => {
   return createFieldWithPresets({
